Move Authorizer helper methods to auth package
